Revert "Bug 2046840 - Simplify path processing code r=firefox-svg-reviewers,dholbert" for causing crashtests at SMILAnimationFunction.cpp
This reverts commit 820675b96d3d52c05be5e154e45bffa5be16c542.

UltraBlame original commit: 15fc84f6d16911cbcd02e01a4f0ff4cbc40994e3

+++ dom/svg/SVGPathData.h
@@ -67,7 +67,7 @@ class SVGPathData {
 
   static already_AddRefed<dom::SVGPathSegment> GetPathSegmentAtLength(
       dom::SVGPathElement* aPathElement, Span<const StylePathCommand> aPath,
-      double aDistance);
+      float aDistance);
 
   void GetMarkerPositioningData(float aZoom, nsTArray<SVGMark>* aMarks) const;
